requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
chardet==5.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
fonttools==4.53.0
fpdf==1.7.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
imgkit==1.2.3
importlib-resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.8.0.76
orjson==3.9.5
packaging==24.1
pandas==2.0.3
pillow==10.3.0
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pyparsing==3.1.2
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.2.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.7.1
tzdata==2024.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.19.2
